requirements.txt
# pip install "apache-airflow[celery]==2.7.0" --constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.7.0/constraints-3.8.txt"
apache-airflow==2.7.0
python-dotenv==1.0.0
pandas==2.0.3
pytest==7.4.0